+++ map_view/display-device-location-with-autopan-modes/gradlew.bat
