+++ map_view/display-device-location-with-autopan-modes/settings.gradle
