+++ map_view/display-device-location-with-autopan-modes/src/main/java/com/esri/samples/display_device_location_with_autopan_modes/DisplayDeviceLocationWithAutopanModesLauncher.java
@@ -1,0 +1,11 @@
+package com.esri.samples.display_device_location_with_autopan_modes;
+
+/**
+   * Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
+   */
+  public class DisplayDeviceLocationWithAutopanModesLauncher {
+
+    public static void main(String[] args) {
+      DisplayDeviceLocationWithAutopanModesSample.main(args);
+    }
+  }
