+++ map_view/display-device-location-with-autopan-modes/src/main/resources/display_device_location/style.css
