--- map_view/display-device-location/src/main/java/com/esri/samples/display_device_location/DisplayDeviceLocationLauncher.java
@@ -1,11 +1,0 @@
-package com.esri.samples.display_device_location;
-
-/**
-   * Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
-   */
-  public class DisplayDeviceLocationLauncher {
-
-    public static void main(String[] args) {
-      DisplayDeviceLocationSample.main(args);
-    }
-  }
